+++ backend/prisma/migrations/20251124104202_remove_employee_email_password/migration.sql
@@ -1,3 +1,0 @@
--- AlterTable
-ALTER TABLE "users" ALTER COLUMN "email" DROP NOT NULL,
-ALTER COLUMN "password" DROP NOT NULL;